Rename translation simba impl file.
